backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic>=2.6.3,<3.0.0
pydantic-settings==2.1.0
email-validator>=2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Agent Framework
pyautogen==0.2.10
openai==1.10.0
together==1.3.0

# Vector Store
chromadb==0.4.22

# Task Queue
celery==5.3.6
redis==5.0.1

# Document Processing
pdfplumber>=0.10.0
pdf2image>=1.16.0
python-docx==1.1.0
pytesseract>=0.3.10  # Optional: enables OCR for scanned PDFs (requires Tesseract installed)
Pillow>=10.0.0  # Required for OCR image processing

# PDF Generation
weasyprint>=60.0
jinja2>=3.1.0

# Web Scraping & Automation
playwright==1.41.0
beautifulsoup4==4.12.3
httpx==0.26.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
PyJWT==2.8.0

# Storage
aioboto3==12.3.0
boto3==1.34.0

# Payments
stripe==7.10.0

# Logging
structlog==24.1.0

# Utilities
tenacity==8.2.3
python-multipart==0.0.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
ruff==0.1.13
mypy==1.8.0
